update readme, clone, config, export and init enhancements, more settings, file re-org, lib/snips makes this requireable, reorg translators

+++ bin/snip-init.js
@@ -5,30 +5,48 @@ var path = require( 'path' );
 var fs = require( 'fs' );
 var shell = require( 'shelljs' );
 var program = require( 'commander' );
-var logger = require( '../lib/logger' );
+var snips = require( '..' );
 
 program
-	.version( require( '../package.json' ).version )
-	.option( '-f, --force', 'Remove ~/.snippets directory first' )
+	.description( 'Initialize a new snippets directory to ~/.snippets' )
+	.option( '-f, --force', 'remove ~/.snippets directory first' )
+	.option( '-b, --blank', 'initialize the directory with no contents' )
+	.option( '-c, --clone [git url]', 'specify a repo to clone into your snippets directory' )
+	.option( '-s, --sample', 'include the sample directory with the clone' )
 	.parse( process.argv );
 
-var goodToInit = true;
 var snippets = path.join( process.env.HOME, '.snippets' );
 
-if ( fs.existsSync( snippets ) ) {
-	if ( program.force ) {
-		shell.rm( '-rf', snippets );
-	} else {
-		goodToInit = false;
-		logger.user( ( '~/.snippets directory already exists' ).blue );
-		logger.user( ( 'snip init --force # to delete ~/.snippets and start fresh' ).grey );
-	}
+var snippetsExists = fs.existsSync( snippets );
+
+if ( snippetsExists && !program.force ) {
+	snips.logger.user( ( '~/.snippets directory already exists' ).red );
+	program.help();
+} else {
+	snips.logger.user( ( 'removed ~/.snippets' ).green );
+	shell.rm( '-rf', snippets );
+	snippetsExists = false;
 }
 
-if ( goodToInit ) {
-	mkpath.sync( snippets );
+mkpath.sync( snippets );
+snips.logger.user( ( 'created ~/.snippets' ).green );
+
+// In the program.clone case
+// only copy sample snippets
+// if the sample flag is passed
+copySample( program.clone );
+
+snips.logger.user( ( 'Done.%s' ).green, program.clone ? ' Starting clone.' : '' );
+
+if ( program.clone ) snips.cli( 'clone', [ program.clone ] );
+
+/******************************************************************************
+ *****     HELPERS    *********************************************************
+ ******************************************************************************/
+function copySample( requireSampleFlag ) {
+	if ( program.blank || ( requireSampleFlag && !program.sample ) ) return;
 	shell.config.silent = true;
 	shell.cp( '-r', path.join( __dirname, '../snippets-sample/' ), path.join( snippets, 'sample/' ) );
 	shell.config.silent = false;
-	logger.user( ( '~/.snippets directory initialized' ).green );
+	snips.logger.user( ( 'created ~/.snippets/sample/' ).green );
 }
